Derp:
A simple, undefined reply when an ignorant comment or action is made.
Brought to life in the South Park series, when Mr. Derp made a guest
apperance at South Park Elementary as the chef for a day, followed by
hitting himself in the head with a hammer and exclaiming "Derp!"

Signed-off-by: scotthartbti <scott.hart.bti@gmail.com>